Added simple password protection

+++ Buttons/Controllers/PrinterController.cs
@@ -11,17 +11,20 @@ namespace Buttons.Controllers
         private readonly ButtonContext context;
         private readonly ILogger<ButtonController> logger;
         private readonly Configuration configuration;
+        private readonly PasswordManager passwordManager;
 
         private Session Session => new(HttpContext.Session, context);
 
         public PrinterController(
             ButtonContext context,
             ILogger<ButtonController> logger,
-            Configuration configuration)
+            Configuration configuration,
+            PasswordManager passwordManager)
         {
             this.context = context;
             this.logger = logger;
             this.configuration = configuration;
+            this.passwordManager = passwordManager;
         }
 
         private AdminButtonViewModel CreateViewModel(Button button) =>
@@ -49,8 +52,28 @@ namespace Buttons.Controllers
             return RedirectToAction(nameof(Index));
         }
 
+        private bool HasAccess
+        {
+            get
+            {
+                int? sessionAccessVersion = Session.GetAdminAccessVersion();
+                return sessionAccessVersion != null &&
+                    sessionAccessVersion == passwordManager.CurrentAccessVersion;
+            }
+        }
+
         public IActionResult Index()
         {
+            if (!passwordManager.HasPassword)
+            {
+                return RedirectToAction(nameof(SetPassword));
+            }
+
+            if (!HasAccess)
+            {
+                return RedirectToAction(nameof(Login));
+            }
+
             var buttons = context.Buttons
                 .Include(b => b.Owner)
                 .Where(b => b.Status >= ButtonStatus.Confirmed)
@@ -64,6 +87,11 @@ namespace Buttons.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Print(int[] buttons, IFormCollection collection)
         {
+            if (!HasAccess)
+            {
+                return RedirectToAction(nameof(Login));
+            }
+
             if (collection.ContainsKey("delete-selection"))
             {
                 return await DeleteButtons(buttons);
@@ -105,6 +133,79 @@ namespace Buttons.Controllers
 
             var buttonViewModels = buttonEntities.Select(CreateViewModel).ToList();
             return View(new AdminButtonListViewModel(buttonViewModels));
+        }
+
+        public IActionResult SetPassword()
+        {
+            if (passwordManager.HasPassword)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> SetPassword(string password)
+        {
+            if (passwordManager.HasPassword)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            var (success, accessVersion) = await passwordManager.SetPasswordAsync(password);
+
+            if (!success)
+            {
+                return RedirectToAction(nameof(SetPassword));
+            }
+
+            logger.LogInformation("Password was set by {}", HttpContext.Connection.RemoteIpAddress);
+            Session.SetAdminAccessVersion(accessVersion);
+            return RedirectToAction(nameof(Index));
+        }
+
+        public IActionResult Login()
+        {
+            if (!passwordManager.HasPassword)
+            {
+                return RedirectToAction(nameof(SetPassword));
+            }
+
+            if (HasAccess)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            return View(new LoginViewModel());
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Login(string password)
+        {
+            if (!passwordManager.HasPassword)
+            {
+                return RedirectToAction(nameof(SetPassword));
+            }
+
+            if (HasAccess)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            var (success, accessVersion) = await passwordManager.CheckPasswordAsync(password);
+
+            if (!success)
+            {
+                logger.LogWarning("Failed login attempt from {}", HttpContext.Connection.RemoteIpAddress);
+                return View(new LoginViewModel("Entered password is invalid."));
+            }
+
+            logger.LogInformation("Successful login from {}", HttpContext.Connection.RemoteIpAddress);
+            Session.SetAdminAccessVersion(accessVersion);
+            return RedirectToAction(nameof(Index));
         }
     }
 }
